[•] Update AyiinUbot Version 0.1.1

+++ AyiinXd/purge.py
@@ -6,7 +6,7 @@
 # <https://www.github.com/AyiinXd/AyiinUbot/blob/main/LICENSE/>.
 #
 # FROM AyiinUbot <https://github.com/AyiinXd/AyiinUbot>
-# t.me/AyiinChat & t.me/AyiinSupport
+# t.me/AyiinChats & t.me/AyiinChannel
 
 
 # ========================×========================
@@ -15,23 +15,25 @@
 
 from fipper import Client
 from fipper.types import Message
-from time import sleep
 
 from pyAyiin import Ayiin, CMD_HELP
-from pyAyiin.pyrogram import eor
+from pyAyiin.pyrogram import eod, eor
+
+from . import yins
 
 
-@Ayiin(["purge"])
-async def purges(client: Client, msg: Message):
-    xx = await eor(msg, "**Mulai Menghapus...**")
-    msg_ids = []
-    count_del = 0
-    if msg.reply_to_message:
-        for start_delete in range(
-            msg.reply_to_message.id,
-            msg.id
-        ):
-            msg_ids.append(start_delete)
+@Ayiin(["purge"], langs=True)
+async def purges(client: Client, msg: Message, _):
+    xx = await eor(msg, _["p"])
+    if await yins.CheckAdmin(client, msg):
+        msg_ids = []
+        count_del = 0
+        if msg.reply_to_message:
+            for start_delete in range(
+                msg.reply_to_message.id,
+                msg.id
+            ):
+                msg_ids.append(start_delete)
             if len(msg_ids) == 100:
                 await client.delete_messages(
                     chat_id=msg.chat.id, message_ids=msg_ids, revoke=True
@@ -42,10 +44,9 @@ async def purges(client: Client, msg: Message):
                 chat_id=msg.chat.id, message_ids=msg_ids, revoke=True
             )
             count_del += len(msg_ids)
-
-    await xx.edit(f"**Berhasil Menghapus** `{count_del}` **pesan.**")
-    sleep(3)
-    await msg.delete()
+        return await xx.edit(_['purge'].format(count_del))
+    else:
+        return await eod(msg, _["admin_5"])
 
 
 CMD_HELP.update(
